Cox regression with simple additive model for Nested Case-Control NCC study

Cox regression with simple additive model for Nested Case-Control NCC study

wrapper_ncc_cox_regression_core_prognostic(
  data,
  tte_var,
  censor_var,
  covariate_vars,
  ncc_vars,
  samplestat_var,
  m,
  match.var,
  match.int,
  return_vars = NULL,
  variable_names = NULL,
  caption = NULL,
  print_pvalues = TRUE,
  print_adjpvalues = TRUE
)

Arguments

data

Data frame preprocessed for the NCC analysis with 'multipleNCC::wpl()'.

tte_var

Name of the time-to-event variable. This variable must be numeric.

censor_var

Name of the censor variable. It has to be numeric and encode 1 for event and 0 for censor.

covariate_vars

Vector with names of covariates that are included in the formula of the simple additive model.

ncc_vars

Vector of names of covariates that were measured in the NCC.

samplestat_var

Name of variable indicating samplestat values. See 'multipleNCC::wpl()'.

m

See 'multipleNCC::wpl()'.

match.var

See 'multipleNCC::wpl()'. It has to be a matrix of continuous values.

match.int

See 'multipleNCC::wpl()'.

return_vars

Vector with names of covariates for which the statistics should be returned. If NULL, statistics are returned for all covariates.

variable_names

Named vector with variable names. If not supplied, variable names are created by replacing in column names underscores with spaces.

caption

Caption for the table with results.

print_pvalues

Logical. Whether to print p-values.

Details

If for a factor covariate that should be returned the reference level has zero count, results are set to NAs because this levels is not used as a reference which means that it is not possible to fit the model that we want.