Tile plot showing co-occurrence and mutual exclusivity of variant types for selected genes

Tile plot showing co-occurrence and mutual exclusivity of variant types for selected genes

wrapper_tileplot(
  data_variants_spread,
  var_samples,
  vars_cat,
  names_cat,
  vars_num = NULL,
  names_num = NULL,
  cut_points = NULL,
  cut_colors = NULL,
  levels_variants,
  colors_variants,
  title = ""
)

Arguments

data_variants_spread
Data frame where each row corresponds to one sample.

var_samples
Column name defining samples.

vars_cat
Column names of categorical variables defining variants for genes of interest. Order of the variables defined order in which they will be plotted starting from the top of the figure.

names_cat
Names corresponding to the categorical variables that should be displayed on the plot, for example, gene names.

vars_num
Column names of numerical variables, for example, one corresponding to TMB.

names_num
Names corresponding to the numerical variables that should be displayed on the plot.

cut_points
List of the same length as vars_num specifying vectors of cutting points for each of the numerical variables in vars_num.

cut_colors
List of the same length as vars_num specifying vectors of colors for the intervals obtained by cutting data using cut_points.

levels_variants
Vector defining the order in which variants should be plotted.

colors_variants
Vector defining colors for the variants.

title
String defining the plot title.