Bclass object

A Bclass object is a container for results generated by the wrappers.

bkable(x, ...)

bforest(x, ...)

bresults(x, ...)

# S4 method for Bclass
bresults(x)

# S4 method for `NULL`
bresults(x)

bresults(x) <- value

# S4 method for Bclass
bresults(x) <- value

boutput(x, ...)

# S4 method for Bclass
boutput(x)

# S4 method for `NULL`
boutput(x)

boutput(x) <- value

# S4 method for Bclass
boutput(x) <- value

bcaption(x, ...)

# S4 method for Bclass
bcaption(x)

# S4 method for `NULL`
bcaption(x)

bcaption(x) <- value

# S4 method for Bclass
bcaption(x) <- value

bheader(x, ...)

# S4 method for Bclass
bheader(x)

# S4 method for `NULL`
bheader(x)

bheader(x) <- value

# S4 method for Bclass
bheader(x) <- value

# S4 method for Bclass
dim(x)

# S4 method for Bclass
nrow(x)

# S4 method for Bclass
ncol(x)

# S4 method for Bclass
colnames(x)

# S4 method for Bclass
colnames(x) <- value

# S4 method for Bclass
rownames(x)

# S4 method for Bclass
rownames(x) <- value

# S4 method for Bclass,ANY
[[(x, i, j)

# S4 method for Bclass
$(x, name)

# S4 method for Bclass,ANY,ANY,ANY
[(x, i, j)

# S4 method for Bclass
rbind(..., deparse.level = 1)

# S4 method for Bclass
cbind(..., deparse.level = 1)

# S4 method for BclassCharacteristics
bkable(x, caption = NULL, header = NULL, font_size = NULL, full_width = NULL)

# S4 method for BclassDE
bkable(x, caption = NULL, header = NULL, font_size = NULL, full_width = NULL)

# S4 method for BclassTesting
bkable(
  x,
  caption = NULL,
  header = NULL,
  block_vars = NULL,
  font_size = NULL,
  full_width = NULL
)

# S4 method for BclassTesting
bforest(
  x,
  mean_var = NULL,
  lower_var = NULL,
  upper_var = NULL,
  block_vars = NULL,
  xlab = NULL,
  clip = NULL,
  xticks = NULL,
  xticks_by = NULL,
  lineheight = "auto"
)

Arguments

value
New values.

object, x
Bclass object.

Details

Methods:

bresults(x): Get a data frame with results.

boutput(x): Get a data frame with output that is dispalyed with kable.

bcaption(x): Get a string with caption that is displayed with kable.

bheader(x): Get a vector with header that is displayed with kable.

bkable(x): Display output with kable.

bforest(x): Generate a forest plot.

Inheriting classes:

Bclass

Slots

results

Data frame with results.

output

Data frame with output that is dispalyed with kable.

caption

String with caption that is displayed in the table.

header

Names integer vector with header that is displayed in the table.

Author

Malgorzata Nowicka