Generate colors for ComplexHeatmap for numerical variables

Generate colors for ComplexHeatmap for numerical variables

format_colors_num(
  x,
  centered = TRUE,
  palette = NULL,
  rev = FALSE,
  trim_values = NULL,
  trim_prop = NULL,
  trim_range = NULL,
  ceiling = FALSE
)

Arguments

x
Vector of numerical values for which we want to specify colors.

centered
Logical. Default value TRUE.

palette
Vector of at least two colors used to create a color palette with 'colorRampPalette' or name of a RColorBrewer palette e.g. "Oranges", "Spectral".

Value

Vector of unique colors for all unique values of x.

Examples

x <- rnorm(20)

palette <- "Spectral"

format_colors_num(x, trim_values = 2.5)

#> function (x = NULL, return_rgb = FALSE, max_value = 1) 
#> {
#>     if (is.null(x)) {
#>         stop_wrap("Please specify `x`\n")
#>     }
#>     att = attributes(x)
#>     if (is.data.frame(x)) 
#>         x = as.matrix(x)
#>     l_na = is.na(x)
#>     if (all(l_na)) {
#>         return(rep(NA, length(l_na)))
#>     }
#>     x2 = x[!l_na]
#>     x2 = ifelse(x2 < breaks[1], breaks[1], ifelse(x2 > breaks[length(breaks)], 
#>         breaks[length(breaks)], x2))
#>     ibin = .bincode(x2, breaks, right = TRUE, include.lowest = TRUE)
#>     res_col = character(length(x2))
#>     for (i in unique(ibin)) {
#>         l = ibin == i
#>         res_col[l] = .get_color(x2[l], breaks[i], breaks[i + 
#>             1], colors[i, ], colors[i + 1, ], space = space)
#>     }
#>     res_col = paste(res_col, transparency_str[1], sep = "")
#>     if (return_rgb) {
#>         res_col = t(col2rgb(as.vector(res_col), alpha = TRUE)/255)
#>         return(res_col)
#>     }
#>     else {
#>         res_col2 = character(length(x))
#>         res_col2[l_na] = NA
#>         res_col2[!l_na] = res_col
#>         attributes(res_col2) = att
#>         return(res_col2)
#>     }
#> }
#> <bytecode: 0x7fcfaf4c5bc8>
#> <environment: 0x7fcfaf4cd218>
#> attr(,"breaks")
#>  [1] -2.5000000 -2.2058824 -1.9117647 -1.6176471 -1.3235294 -1.0294118
#>  [7] -0.7352941 -0.4411765 -0.1470588  0.1470588  0.4411765  0.7352941
#> [13]  1.0294118  1.3235294  1.6176471  1.9117647  2.2058824  2.5000000
#> attr(,"colors")
#>              red     green        blue
#>  [1,] 0.03137255 0.1882353 0.419607843
#>  [2,] 0.03137255 0.3176471 0.611764706
#>  [3,] 0.12941176 0.4431373 0.709803922
#>  [4,] 0.25882353 0.5725490 0.776470588
#>  [5,] 0.41960784 0.6823529 0.839215686
#>  [6,] 0.61960784 0.7921569 0.882352941
#>  [7,] 0.77647059 0.8588235 0.937254902
#>  [8,] 0.87058824 0.9215686 0.968627451
#>  [9,] 0.96862745 0.9843137 1.000000000
#> [10,] 1.00000000 0.9607843 0.921568627
#> [11,] 0.99607843 0.9019608 0.807843137
#> [12,] 0.99215686 0.8156863 0.635294118
#> [13,] 0.99215686 0.6823529 0.419607843
#> [14,] 0.99215686 0.5529412 0.235294118
#> [15,] 0.94509804 0.4117647 0.074509804
#> [16,] 0.85098039 0.2823529 0.003921569
#> [17,] 0.65098039 0.2117647 0.011764706
#> [18,] 0.49803922 0.1529412 0.015686275
#> attr(,"transparency")
#> [1] 0
#> attr(,"space")
#> [1] "LAB"